+++ lib/ts/recipe/emailpassword/types.ts
@@ -131,6 +131,13 @@ export type RecipeInterface = {
         userContext: UserContext;
     }): Promise<{ status: "OK"; user: User; recipeUserId: RecipeUserId } | { status: "WRONG_CREDENTIALS_ERROR" }>;
 
+    verifyCredentials(input: {
+        email: string;
+        password: string;
+        tenantId: string;
+        userContext: UserContext;
+    }): Promise<{ status: "OK" } | { status: "WRONG_CREDENTIALS_ERROR" }>;
+
     /**
      * We pass in the email as well to this function cause the input userId
      * may not be associated with an emailpassword account. In this case, we