+++ lib/ts/recipe/thirdpartyemailpassword/types.ts
@@ -201,6 +201,18 @@ export type RecipeInterface = {
         | { status: "WRONG_CREDENTIALS_ERROR" }
     >;
 
+    emailPasswordVerifyCredentials(input: {
+        email: string;
+        password: string;
+        tenantId: string;
+        userContext: UserContext;
+    }): Promise<
+        | {
+              status: "OK";
+          }
+        | { status: "WRONG_CREDENTIALS_ERROR" }
+    >;
+
     createResetPasswordToken(input: {
         userId: string;
         email: string;